Redirect on signup to verification page
